discord-reset-bot/requirements.txt
aiohttp>=3.7.4
discord.py>=1.6.0
python-dotenv>=0.15.0